+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added safety check for multiple `Model.reindex`
 - Added `deep_paging` option
+- Added request parameters to `curl` representation
 
 ## 4.1.1 (2019-11-19)